requirements-base.txt
mq-transport[sqs]@git+https://github.com/Bondifuzz/mq-transport.git@0.1.0
aiohttp==3.7.4
aioarangodb==0.1.2
coloredlogs==15.0
prometheus-client==0.11.0
pydantic[email]==1.8.2
PyYAML==5.4.1
jira==3.1.1